Crate wasm3

A rust wrapper for WASM3.

Re-exports

pub use ffi as wasm3_sys;

Modules

error

Error related functionality of wasm3.

Macros

make_func_wrapper

A convenience macro for creating a wrapper function that can be linked to wasm3.

Structs

CallContext

Calling Context for a host function.

Environment

An environment is required to construct Runtimes from.

Function

A callable wasm3 function. This has a generic call function for up to 26 parameters emulating an overloading behaviour without having to resort to tuples. These are hidden to not pollute the documentation.

Module

A loaded module belonging to a specific runtime. Allows for linking and looking up functions.

ParsedModule

A parsed module which can be loaded into a Runtime.

Runtime

A runtime context for wasm3 modules.

Traits

WasmArg

Tait implemented by types that can be passed to wasm.

WasmArgs

Helper tait implemented by tuples to emulate “variadic generics”.

WasmType

Trait implemented by types that can be passed to and from wasm.

Type Definitions

RawCall

Type of a raw host function for wasm3.